Refactoring LaunchFileReader
